pyproject.toml
[project]
name = "brief-ai-vs-software-engineering"
version = "0.1.0"
description = "Build a concrete and critical brief comparing Software Engineering vs AI Engineering."
requires-python = ">=3.10"
dependencies = [
  "openai>=1.60.0",
  "langchain>=0.3.0",
  "langchain-openai>=0.2.0",
  "langgraph>=0.2.0",
  "graphviz>=0.20.1",
  "pillow>=10.0.0",
  "python-dotenv>=1.0.1",
  "nbformat>=5.10.4",
  "nbclient>=0.10.2",
  "ipykernel>=6.29.5",
  "pytest>=8.3.0",
  "pytest-mock>=3.12.0",
  "pytest-cov>=4.1.0",
  "pydantic>=2.0.0",
  "chromadb>=0.5.0",
  "langchain-community>=0.3.0",
  "langchain-experimental>=0.3.0",
  "scikit-learn>=1.3.0",
  "matplotlib>=3.8.0",
  "numpy>=1.26.0",
  "pandas>=2.1.0",
  "langchain-core>=1.2.9",
  "langchain-tavily>=0.2.0",
  "tavily-python>=0.7.0",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "integration: integration tests that may call external APIs",
    "unit: fast unit tests",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"
